INITIAL COMMIT
This is the first commit of the Pong game project.

This project will have a broader read me in future and a broader description with goals for the project, but broadly it is Pong with a simple physics engine and dash ability for high octane ponging. I have a design doc written out
physically, I just have to actually port it into here.

This commit specifically adds the skeleton of that 2D physics engine, which though it may not be insane, I am deeply proud of. It has 3 vectors for each physics object for position, velocity, and acceleration and essentially uses Euler
integration to apply acceleration to velocity, which is then applied to the position. There are specific game objects for the ball and paddle, which helps encapsulate the logic. Interfaces are used to integrate renderable and physics
code into each object, and then that code is used in the game screen methods.

The only other notable thing here is that this is also my first foray into making genuinely usable interfaces, as well as different objects and classes within the LibGDX structure. Overall, this project has been a massive learning
experience already, and I think the 2-week timeframe I set out will be absolutely more than enough (I say before spending another month on this).