ENGINE ADDITIONS
This branch will be all about adding additions to the 2D Physics Engine. This commit adds a few important things.

1) Drag forces on the paddles, so pressing them does not give them the same acceleration for the whole game, but starts and stops them as if they have a thruster that turns on and off with each button press.
2) starting horizontal velocity for the ball, so it begins the game by moving towards the opponent who reflects it.
3) Bounce functionality for the ball when it hits the ceiling
4) Bounce functionality for the ball when it hits a paddle, reflecting its angle

STILL NEED TO ADD:
Changes in the reflected ball angle based on the paddle's velocity vector. This will be done by passing a vector into the bounce method and taking the y component's sign (less than or greater than zero), and if less than then add the
y component in that direction, if positive add it in the other direction.
ORGANIZATION of code to their individual classes, as commented in the project.
DASH for the paddles