ENGINE ADDITIONS 2: Working Pong, Bug fixes, Score
This commit fixes a few bugs that were present, namely with the mixing up of delta in the enforceTopSpeed code and the bounce code. Now, all these behind-the-scenes methods are using velocity in units per second, whereas the actual
move takes those velocities and translates based on delta. Now, the ball no longer randomly changes velocities or randomly drops deeply into low velocities*. The ball also now properly bounces, and the velocity of the paddle is added
properly to the ball; meaning if the paddle is moving down, the ball bounces down, and if the paddle is moving up, the ball bounces up (or at least gets velocity added in the direction of the paddle movement).

This commit also adds score counting which is working properly; when the ball goes beyond the world border horizontally, score is added to the player or the enemy side.

Overall, this commit brings this project to a minimally viable working Pong program! At some point in the future I will change the name, but for now this is fine (when I publish it). Eventually I'll also make a make file.

*(except for some edge-cases, which hopefully should be fixed; when the ball is reset at a very similar frame to when it is hit with a paddle, that may transfer its x and y velocities and make its starting velocity low. This should be
fixed in the code now by making it so the ball is always moved towards its top speed, if it is below its top speed acceleration is added, and if it is above its top speed, acceleration in the opposite direction is added. Both are of
magnitude equal to the velocity. This should keep the ball at a roughly constant speed, with some acceleration beyond it but that always gets brought back to its 'resting place').