ENGINE ADDITIONS (#1)
* ENGINE ADDITIONS

This branch will be all about adding additions to the 2D Physics Engine. This commit adds a few important things.

1) Drag forces on the paddles, so pressing them does not give them the same acceleration for the whole game, but starts and stops them as if they have a thruster that turns on and off with each button press.
2) starting horizontal velocity for the ball, so it begins the game by moving towards the opponent who reflects it.
3) Bounce functionality for the ball when it hits the ceiling
4) Bounce functionality for the ball when it hits a paddle, reflecting its angle

STILL NEED TO ADD:
Changes in the reflected ball angle based on the paddle's velocity vector. This will be done by passing a vector into the bounce method and taking the y component's sign (less than or greater than zero), and if less than then add the
y component in that direction, if positive add it in the other direction.
ORGANIZATION of code to their individual classes, as commented in the project.
DASH for the paddles

* ENGINE ADDITIONS 2: Working Pong, Bug fixes, Score

This commit fixes a few bugs that were present, namely with the mixing up of delta in the enforceTopSpeed code and the bounce code. Now, all these behind-the-scenes methods are using velocity in units per second, whereas the actual
move takes those velocities and translates based on delta. Now, the ball no longer randomly changes velocities or randomly drops deeply into low velocities*. The ball also now properly bounces, and the velocity of the paddle is added
properly to the ball; meaning if the paddle is moving down, the ball bounces down, and if the paddle is moving up, the ball bounces up (or at least gets velocity added in the direction of the paddle movement).

This commit also adds score counting which is working properly; when the ball goes beyond the world border horizontally, score is added to the player or the enemy side.

Overall, this commit brings this project to a minimally viable working Pong program! At some point in the future I will change the name, but for now this is fine (when I publish it). Eventually I'll also make a make file.

*(except for some edge-cases, which hopefully should be fixed; when the ball is reset at a very similar frame to when it is hit with a paddle, that may transfer its x and y velocities and make its starting velocity low. This should be
fixed in the code now by making it so the ball is always moved towards its top speed, if it is below its top speed acceleration is added, and if it is above its top speed, acceleration in the opposite direction is added. Both are of
magnitude equal to the velocity. This should keep the ball at a roughly constant speed, with some acceleration beyond it but that always gets brought back to its 'resting place').

* ENGINE ADDITIONS 3: REORGANIZATION OF CODE

This commit mainly just re-organizes code in preparation for adding new features. Some things can still be re-organized, but I have to figure out how to do it without breaking things (namely, the check collision stuff that's in the
logic method right now. That can definitely be moved into the ball class, but I need to figure out how to get it to actually work properly while there, since when I did it last things seemed to break).

This just moves things out of game screen to make that code cleaner, and makes things more encapsulated into their proper classes. With this, I should be ready to merge this branch into main and then work on the dash ability.