ENGINE ADDITIONS 3: REORGANIZATION OF CODE
This commit mainly just re-organizes code in preparation for adding new features. Some things can still be re-organized, but I have to figure out how to do it without breaking things (namely, the check collision stuff that's in the
logic method right now. That can definitely be moved into the ball class, but I need to figure out how to get it to actually work properly while there, since when I did it last things seemed to break).

This just moves things out of game screen to make that code cleaner, and makes things more encapsulated into their proper classes. With this, I should be ready to merge this branch into main and then work on the dash ability.