FINAL COMMIT FOR 1.0.0
After a hiatus where I worked on other projects in the background (and got busy with life), this will be the last commit for 1.0.0. I will now be packing it and publishing it on itch.io.

Added:
1. The dash mechanic! It's still a little wonky, but it does generally work.
2. A reset mechanic, so you can go back to the title screen at any time.
3. Instructions on the title screen for all the controls.

In the future, I want to of course refine the systems in place, make them better, and make them feel more like a game, and I also want to add bells and whistles. I.E. sounds, title screen improvements, options, etc.
But for now, this works as an MVP.